backend/pyproject.toml
[tool.poetry]
name = "marketplace-backend"
version = "0.1.0"
description = ""
authors = ["vllscarf <xxxisa2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
django-mptt = "^0.16.0"
django-filter = "^24.3"
django-treebeard = "^4.7.1"
django-js-asset = "^2.2.0"
djangorestframework-simplejwt = "^5.3.1"
rules = "^3.4"
django-cors-headers = "^4.4.0"
django-urls = "^1.1.3"
django-extensions = "^3.2.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
